Refactor storage layer into domain repositories
