+++ server/storage/products.ts
@@ -1,0 +1,40 @@
+import {
+  products,
+  type Product,
+  type InsertProduct,
+} from "@shared/schema";
+import { db } from "../db";
+import { eq } from "drizzle-orm";
+
+export class ProductsRepository {
+  async getProducts(): Promise<Product[]> {
+    return await db
+      .select()
+      .from(products)
+      .where(eq(products.isActive, true))
+      .orderBy(products.createdAt);
+  }
+
+  async getProduct(id: string): Promise<Product | undefined> {
+    const [product] = await db.select().from(products).where(eq(products.id, id));
+    return product;
+  }
+
+  async createProduct(product: InsertProduct): Promise<Product> {
+    const [createdProduct] = await db.insert(products).values(product).returning();
+    return createdProduct;
+  }
+
+  async updateProduct(id: string, product: Partial<InsertProduct>): Promise<Product> {
+    const [updatedProduct] = await db
+      .update(products)
+      .set({ ...product, updatedAt: new Date() })
+      .where(eq(products.id, id))
+      .returning();
+    return updatedProduct;
+  }
+
+  async deleteProduct(id: string): Promise<void> {
+    await db.update(products).set({ isActive: false }).where(eq(products.id, id));
+  }
+}
